backend/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
python-decouple
django-allauth
djangorestframework-simplejwt
django-cors-headers
drf-yasg